INS-464: rename experiment
